+++ main.js
@@ -1,4 +1,4 @@
-var IMVVM = require('./src/core.js');
+var Astarisx = require('./src/core.js');
 require('./lib/custom-event-polyfill');
 
-module.exports = IMVVM;
+module.exports = Astarisx;